% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImVOl.R
\name{ImVoL}
\alias{ImVoL}
\title{Volume Prediction of Trees Using Linear and Nonlinear Allometric Equations}
\usage{
ImVoL(data)
}
\arguments{
\item{data}{Datasets}
}
\value{
\itemize{
  \item results: Results
}
}
\description{
Volume Prediction of Trees Using Linear and Nonlinear Allometric Equations
}
\examples{
library("ImVol")
data <- system.file("extdata", "data_test.csv", package = "ImVol")
Data<- read.csv(data)
Vol <- ImVoL(Data)
}
\references{
\itemize{
\item Sharma, A., Gupta.R.K., Mahajan, P.K and Shilp. 2017.Volume Equation for Cedrus deodara in Kullu District of Himachal Pradesh. Indian Journal of Ecology . 44(6) : 781-784.http://dx.doi.org/10.13140/RG.2.2.33786.62407
\item Tabassum, A., Jeelani, M.I.,Sharma,M., Rather, K R ., Rashid, I and Gul,M.2022.  Predictive Modelling of Height and Diameter Relationships of Himalayan Chir Pine .Agricultural Science Digest - A Research Journal. DOI:10.18805/ag.D-555
}
}
